requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.0=py38h3931269_0
contourpy=1.0.7=pypi_0
cudatoolkit=11.3.1=h9edb442_10
cudnn=8.2.1.32=h86fa8c9_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
et_xmlfile=1.1.0=py38h06a4308_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.2.10=h9c3ff4c_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.39.3=pypi_0
freetype=2.10.4=h0708190_1
future=0.18.3=pyhd8ed1ab_0
giflib=5.2.1=h36c2ea0_2
glib=2.69.1=he621ea3_2
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=hf484d3e_1000
importlib_resources=5.12.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.12.0=pyh41d4057_0
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.2.0=pypi_0
jpeg=9e=h166bdaf_1
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=5.3.0=py38h578d9bd_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.19.3=h3790be6_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.17=h5eee18b_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=12_linux64_mkl
libllvm10=10.0.1=he513fc3_3
libmtl=1.1.5=dev_0
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.18.0=h780b84a_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_2
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h9c3ff4c_1
magma=2.5.4=h6103c52_2
matplotlib=3.7.1=py38h578d9bd_0
matplotlib-base=3.7.1=py38h417a72b_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=pyh9f0ad1d_0
nccl=2.12.12.1=h0800d71_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
ninja=1.11.0=h924138e_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.20.0=pypi_0
numpy-base=1.23.5=py38h31eccc5_0
openpyxl=3.0.10=py38h5eee18b_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=py38h417a72b_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=23.0.1=py38h06a4308_0
platformdirs=3.2.0=pyhd8ed1ab_0
ply=3.11=py_1
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.15.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.9.1=cuda112py38h4f2a933_3
pytorch-gpu=1.9.1=cuda112py38h0bbbad9_3
pytz=2022.7=py38h06a4308_0
pyzmq=19.0.2=py38ha71036d_2
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h9b69904_2
sqlite=3.41.1=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=0.9.1=py38
torchvision=0.9.0=pypi_0
tornado=6.1=py38h0a891b7_3
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py38h06a4308_0
xz=5.2.10=h5eee18b_1
zeromq=4.3.4=h9c3ff4c_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
